+++ NEWS.md
@@ -1,3 +1,8 @@
+# v1.0.1
+
+fixed bug where `fitGLS_partition` would not work properly when `ncores > 1` and
+`formula == formula0`. 
+
 # v1.0.0
 
 `remotePARTS` has been completely restructured from previous development builds.